+++ tests/dummy/app/controllers/index.js
@@ -3,11 +3,27 @@ import { computed } from 'ember-can';
 
 var Post = Ember.Object.extend();
 
+var editablePost = Post.create({
+  author: 42,
+  title: "Something"
+});
+
+var otherPost = Post.create({
+  author: 99,
+  title: "Something Else"
+});
+
 export default Ember.ObjectController.extend({
   ability:      computed.ability("post"),
   canWritePost: Ember.computed.alias("ability.canWrite"),
-  post: Post.create({
-    author: 42,
-    title: "Something"
-  })
+  post: editablePost,
+
+  actions: {
+    selectEditable: function() {
+      this.set("post", editablePost);
+    },
+    selectOther:    function() {
+      this.set("post", otherPost);
+    }
+  }
 });