+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Master
 
+* Observe changes to the resource
+* Fix can-unless helper
+
 ## v0.1.0
 
 * Initial usable version